Clase 05

+++ Clase05/EurekaService/src/pe/egcc/eureka/service/espec/CuentaEspec.java
@@ -2,6 +2,6 @@ package pe.egcc.eureka.service.espec;
 
 public interface CuentaEspec {
   
-  int procRetiro( String cuenta, double importe, String clave, String empleado);
+  double procRetiro( String cuenta, double importe, String clave, String empleado);
   
 }
